Move RuleState into rule directory.
