+++ great_expectations/rule_based_profiler/rule/rule_output.py
@@ -10,7 +10,7 @@ from great_expectations.rule_based_profiler.parameter_container import (
     get_fully_qualified_parameter_names,
     get_parameter_values_for_fully_qualified_parameter_names,
 )
-from great_expectations.rule_based_profiler.rule_state import RuleState
+from great_expectations.rule_based_profiler.rule.rule_state import RuleState
 
 
 class RuleOutput:
